+++ src/metadata/metadata_io.c
@@ -504,7 +504,7 @@ err:
 }
 
 int metadata_io_write_i_asynch(ocf_cache_t cache, ocf_queue_t queue,
-		void *context, uint32_t page, uint32_t count, int flags,
+		void *context, uint64_t page, uint64_t count, int flags,
 		ocf_metadata_io_event_t fill_hndl,
 		ocf_metadata_io_end_t compl_hndl,
 		struct ocf_alock *mio_conc)
@@ -514,7 +514,7 @@ int metadata_io_write_i_asynch(ocf_cache_t cache, ocf_queue_t queue,
 }
 
 int metadata_io_read_i_asynch(ocf_cache_t cache, ocf_queue_t queue,
-		void *context, uint32_t page, uint32_t count, int flags,
+		void *context, uint64_t page, uint64_t count, int flags,
 		ocf_metadata_io_event_t drain_hndl,
 		ocf_metadata_io_end_t compl_hndl)
 {
